refactor: centralize effective config and phase ownership
