+++ src/frame_compare/orchestration/phase_tasks.py
@@ -1,95 +1,53 @@
-"""Concrete orchestration phase work.
-
-The execution module owns phase ordering and timing. This module owns the
-phase bodies and shared translation helpers used by preparation and execution.
-"""
+"""Alignment and render phase work plus local frame-normalization helpers."""
 
 from __future__ import annotations
 
 from dataclasses import dataclass, replace
-from fractions import Fraction
 from pathlib import Path
 from typing import TYPE_CHECKING
 
-import httpx
 import structlog
 
-import frame_compare.analysis.cache_io as cache_io
-from frame_compare.analysis.errors import MetricsCalculationError, SelectionError
+from frame_compare.analysis.errors import SelectionError
 from frame_compare.analysis.frame_plan import create_frame_plan
-from frame_compare.analysis.metrics import calculate_metrics
 from frame_compare.analysis.selection import select_frames
 from frame_compare.analysis.types import (
-    CacheLoadResult,
     FrameMetrics,
-    FrameSelection,
     SelectionBreakdown,
     SelectionDetail,
     SelectionDetailsByFrame,
 )
-from frame_compare.analysis.window import SelectionWindow
-from frame_compare.config.errors import ConfigValidationError
-from frame_compare.config.schema import AnalysisConfig, ConfigSchema, OverlayMode
-from frame_compare.errors import JSONValue
+from frame_compare.config.schema import AnalysisConfig, OverlayMode
 from frame_compare.orchestration.context import (
     ClipAlignmentState,
     ClipProbeSnapshot,
     ClipState,
     RunContext,
 )
-from frame_compare.orchestration.types import (
-    AlignPhaseOutput,
-    AnalyzePhaseOutput,
-    ConfirmSlowpicsUploadPhaseOutput,
-    DoviPhaseOutput,
-    FramePlanPhaseOutput,
-    MetadataPhaseOutput,
-    MetadataPrefetch,
-    PostReportCleanupPhaseOutput,
-    PublishPhaseOutput,
-    RenderArtifacts,
-    RenderPhaseOutput,
-    ReportPhaseOutput,
-    SlowpicsUploadConfirmationFn,
-    SlowpicsUploadConfirmationRequest,
+from frame_compare.orchestration.phase_selection import (
+    build_initial_selection_details_by_source_frame,
+    generated_frame_count,
+    map_aligned_to_source_frame,
+    selection_breakdown_with_source_offset,
+    selection_detail_for_frame,
+    selection_details_with_source_offset,
+    selection_label_for_frame,
+    selection_timecode_for_frame,
+    source_frames_for_reference_base_domain,
+    to_overlay_selection_detail,
 )
+from frame_compare.orchestration.types import AlignPhaseOutput, RenderArtifacts, RenderPhaseOutput
 from frame_compare.render.backend.ffmpeg import FFmpegRunner
 from frame_compare.services.alignment import (
     align_clips,
     calculate_alignment_trims,
     format_rejected_alignment_warning,
 )
-from frame_compare.services.errors import AudioAlignmentError, SlowpicsError
-from frame_compare.services.metadata import resolve_metadata
-from frame_compare.services.publishers import publish_to_slowpics
-from frame_compare.services.report.display import (
-    SourceFrameSelectionDetail,
-    frame_detail_for_source_frame,
-)
-from frame_compare.services.report.entry import generate_report
-from frame_compare.services.report.payload import (
-    FrameDetail,
-    ReportData,
-    clip_info_from_state,
-)
-from frame_compare.services.slowpics_post_upload import (
-    SlowpicsPostUploadRequest,
-    run_slowpics_post_upload_actions,
-)
-from frame_compare.services.slowpics_upload_plan import (
-    SlowpicsUploadClip,
-    build_slowpics_upload_plan,
-)
-from frame_compare.services.types import AlignmentConfig, MetadataConfig, TmdbMetadata
-from frame_compare.utils.cache_errors import CacheCorruptionError, CacheVersionMismatchError
-from frame_compare.utils.types import WorkspacePaths
+from frame_compare.services.errors import AudioAlignmentError
+from frame_compare.services.types import AlignmentConfig
 from frame_compare.vs.props import range_label_from_props
 
 log = structlog.get_logger()
-
-REPORT_CONFIRMATION_UNAVAILABLE_WARNING = (
-    "slow.pics upload skipped because report confirmation was unavailable"
-)
 
 if TYPE_CHECKING:
     from frame_compare.render.types import (
@@ -97,430 +55,6 @@ if TYPE_CHECKING:
         OverlayDolbyVisionMetadata,
         OverlayFrameMeasurement,
         OverlaySelectionDetail,
-    )
-    from frame_compare.vs.loader import VSLoader
-
-
-def build_metadata_config(config: ConfigSchema) -> MetadataConfig:
-    """Build metadata service config from run config."""
-    return MetadataConfig(
-        api_key=config.tmdb.api_key,
-        unattended=config.tmdb.unattended,
-        timeout_seconds=config.tmdb.timeout_seconds,
-        year_tolerance=config.tmdb.year_tolerance,
-        category_preference=config.tmdb.category_preference,
-    )
-
-
-async def resolve_run_metadata(
-    *,
-    filenames: list[str],
-    config: ConfigSchema,
-    client: httpx.AsyncClient,
-) -> TmdbMetadata | None:
-    return await resolve_metadata(
-        filenames=filenames,
-        config=build_metadata_config(config),
-        client=client,
-    )
-
-
-def select_initial_frame_plan(ctx: RunContext) -> FramePlanPhaseOutput:
-    window_start, frame_count = _selection_window_for_context(ctx)
-    user_frames = _user_frames_in_window(
-        ctx.config.analysis, ctx.reference, window_start, frame_count
-    )
-    user_frame_set = set(user_frames)
-    dropped_user_frames = sorted(set(ctx.config.analysis.user_frames) - set(user_frames))
-    source_offset = ctx.reference.trim.trim_start_frames + window_start
-    selectable_random_indices = [
-        frame_index
-        for frame_index in range(frame_count)
-        if source_offset + frame_index not in user_frame_set
-    ]
-    random_count = ctx.config.analysis.random_frame_count
-    if random_count > len(selectable_random_indices):
-        raise SelectionError(
-            "insufficient random candidates after user frames",
-            requested=random_count,
-            found=len(selectable_random_indices),
-        )
-    random_frames = [
-        selectable_random_indices[frame] + window_start
-        for frame in create_frame_plan(
-            num_frames=len(selectable_random_indices),
-            count=random_count,
-            seed=ctx.config.analysis.random_seed,
-        ).frames
-    ]
-    random_source_frames = [ctx.reference.trim.trim_start_frames + frame for frame in random_frames]
-    selected_frames = sorted(
-        {
-            *(frame - ctx.reference.trim.trim_start_frames for frame in user_frames),
-            *random_frames,
-        }
-    )
-    requested_initial_count = len(ctx.config.analysis.user_frames) + (
-        ctx.config.analysis.random_frame_count
-    )
-    if requested_initial_count > 0 and not selected_frames:
-        raise SelectionError(
-            "no selectable user or random frames remain after trims/windowing",
-            requested=requested_initial_count,
-            found=0,
-        )
-    warnings = (
-        [
-            "frame selection: dropped user frame(s) outside trims/windowing: "
-            + ", ".join(str(frame) for frame in dropped_user_frames)
-        ]
-        if dropped_user_frames
-        else []
-    )
-    return FramePlanPhaseOutput(
-        selected_frames=selected_frames,
-        selection_breakdown=SelectionBreakdown(user=user_frames, random=random_source_frames),
-        selection_details_by_source_frame=_initial_selection_details_by_source_frame(
-            user_source_frames=user_frames,
-            random_source_frames=random_source_frames,
-            fps=ctx.reference.effective_fps,
-        ),
-        warnings=warnings,
-    )
-
-
-def _initial_selection_details_by_source_frame(
-    *,
-    user_source_frames: list[int],
-    random_source_frames: list[int],
-    fps: Fraction,
-) -> SelectionDetailsByFrame:
-    details: SelectionDetailsByFrame = {}
-    for frame in user_source_frames:
-        details[frame] = SelectionDetail(
-            frame_index=frame,
-            label="User",
-            source="frame_plan",
-            timecode=_selection_timecode_for_frame(frame, fps),
-            clip_role="frame_plan",
-            notes="user",
-        )
-    for frame in random_source_frames:
-        details.setdefault(
-            frame,
-            SelectionDetail(
-                frame_index=frame,
-                label="Random",
-                source="frame_plan",
-                timecode=_selection_timecode_for_frame(frame, fps),
-                clip_role="frame_plan",
-                notes="random",
-            ),
-        )
-    return details
-
-
-def _user_frames_in_window(
-    config: AnalysisConfig,
-    reference: ClipState,
-    window_start: int,
-    frame_count: int,
-) -> list[int]:
-    source_start = reference.trim.trim_start_frames + window_start
-    source_end = source_start + frame_count
-    return sorted({frame for frame in config.user_frames if source_start <= frame < source_end})
-
-
-def _selection_window_for_context(ctx: RunContext) -> tuple[int, int]:
-    window = ctx.selection_window
-    if window.frame_count > 0:
-        return window.start_frame, window.frame_count
-    return 0, ctx.reference.effective_num_frames()
-
-
-def _generated_frame_count(config: AnalysisConfig) -> int:
-    return (
-        config.random_frame_count
-        + config.dark_frame_count
-        + config.bright_frame_count
-        + config.motion_frame_count
-    )
-
-
-def _source_offset_for_reference_window(reference: ClipState, window_start: int) -> int:
-    return reference.trim.trim_start_frames + window_start
-
-
-def run_analyze_phase(
-    ctx: RunContext,
-    *,
-    input_videos: list[Path],
-    workspace: WorkspacePaths,
-    require_cache_only: bool = False,
-    vs_loader: VSLoader | None = None,
-) -> AnalyzePhaseOutput:
-    selection_domain = ctx.analysis_selection_domain
-    fingerprint = cache_io.compute_cache_key(
-        input_videos,
-        ctx.config.analysis,
-        selection_domain=selection_domain,
-    )
-    cache_result = cache_io.load_cached_metrics(workspace.cache_dir, fingerprint, clips=[])
-    metrics_cache_hit = cache_result.success and cache_result.metrics is not None
-    if require_cache_only:
-        metrics = _require_cached_metrics(
-            cache_result=cache_result,
-            cache_dir=workspace.cache_dir,
-            input_videos=input_videos,
-            fingerprint=fingerprint,
-        )
-    else:
-        metrics = calculate_metrics(
-            video_paths=input_videos,
-            config=ctx.config.analysis,
-            cache_dir=workspace.cache_dir,
-            reporter=ctx.reporter,
-            vs_loader=vs_loader,
-            selection_domain=selection_domain,
-            effective_fps=ctx.reference.effective_fps,
-        )
-    selection = _select_frames_for_selection_domain(
-        metrics=metrics,
-        reference=ctx.reference,
-        selection_window=ctx.selection_window,
-        config=ctx.config.analysis,
-    )
-    return AnalyzePhaseOutput(
-        selected_frames=list(selection.frames),
-        selection_breakdown=selection.breakdown,
-        metrics_cache_hit=metrics_cache_hit,
-        analysis_metrics=metrics,
-        selection_details_by_source_frame=dict(selection.selection_details),
-    )
-
-
-def _require_cached_metrics(
-    *,
-    cache_result: CacheLoadResult,
-    cache_dir: Path,
-    input_videos: list[Path],
-    fingerprint: str,
-) -> FrameMetrics:
-    if cache_result.success and cache_result.metrics is not None:
-        return cache_result.metrics
-
-    cache_path = cache_io.find_metrics_cache_file(cache_dir, fingerprint)
-    expected_cache_path = cache_dir / cache_io.metrics_cache_filename(input_videos, fingerprint)
-    error_cache_path = cache_path or expected_cache_path
-    reason = cache_result.reason
-    if reason == "corrupted":
-        raise CacheCorruptionError(error_cache_path)
-    if reason == "version_mismatch":
-        found = cache_io.read_cache_version(error_cache_path) or "unknown"
-        raise CacheVersionMismatchError(found, str(cache_io.CACHE_VERSION))
-    raise MetricsCalculationError(f"Cached metrics missing or mismatched ({reason}).")
-
-
-def _select_frames_for_selection_domain(
-    *,
-    metrics: FrameMetrics,
-    reference: ClipState,
-    selection_window: SelectionWindow,
-    config: AnalysisConfig,
-) -> FrameSelection:
-    window_start, frame_count = (
-        (selection_window.start_frame, selection_window.frame_count)
-        if selection_window.frame_count > 0
-        else (0, reference.effective_num_frames())
-    )
-    if (
-        reference.trim.trim_start_frames == 0
-        and reference.trim.trim_end_frame_inclusive is None
-        and frame_count == reference.effective_num_frames()
-        and window_start == 0
-    ):
-        return select_frames(
-            metrics=metrics,
-            config=_config_for_selection_window(
-                config=config,
-                reference=reference,
-                window_start=window_start,
-                frame_count=frame_count,
-            ),
-        )
-
-    if frame_count <= 0:
-        raise SelectionError("reference source trims leave no selectable frames", 0, 0)
-
-    trimmed_metrics = _trimmed_metrics_for_overlap(
-        metrics=metrics,
-        trim_start_frame=reference.trim.trim_start_frames + window_start,
-        frame_count=frame_count,
-    )
-    selection = select_frames(
-        metrics=trimmed_metrics,
-        config=_config_for_selection_window(
-            config=config,
-            reference=reference,
-            window_start=window_start,
-            frame_count=frame_count,
-        ),
-    )
-    source_offset = _source_offset_for_reference_window(reference, window_start)
-    return FrameSelection(
-        frames=[frame + window_start for frame in selection.frames],
-        seed=selection.seed,
-        breakdown=_selection_breakdown_with_source_offset(
-            selection.breakdown,
-            source_offset=source_offset,
-        ),
-        selection_details=_selection_details_with_source_offset(
-            dict(selection.selection_details),
-            source_offset=source_offset,
-            fps=trimmed_metrics.metadata.fps,
-        ),
-    )
-
-
-def _config_for_selection_window(
-    *,
-    config: AnalysisConfig,
-    reference: ClipState,
-    window_start: int,
-    frame_count: int,
-) -> AnalysisConfig:
-    source_start = reference.trim.trim_start_frames + window_start
-    source_end = source_start + frame_count
-    local_user_frames = [
-        frame - source_start for frame in config.user_frames if source_start <= frame < source_end
-    ]
-    return config.model_copy(
-        update={
-            "user_frames": local_user_frames,
-        }
-    )
-
-
-def selection_label_for_frame(frame: int, breakdown: SelectionBreakdown | None) -> str | None:
-    if breakdown is None:
-        return None
-    if frame in breakdown.user:
-        return "User"
-    if frame in breakdown.quantile_dark:
-        return "Dark"
-    if frame in breakdown.quantile_bright:
-        return "Bright"
-    if frame in breakdown.motion:
-        return "Motion"
-    if frame in breakdown.random:
-        return "Random"
-    return None
-
-
-def _selection_detail_for_frame(
-    frame: int,
-    details_by_source_frame: dict[int, SelectionDetail] | None,
-) -> SelectionDetail | None:
-    if details_by_source_frame is None:
-        return None
-    return details_by_source_frame.get(frame)
-
-
-def _report_selection_detail(detail: SelectionDetail | None) -> SourceFrameSelectionDetail | None:
-    if detail is None:
-        return None
-    return SourceFrameSelectionDetail(
-        label=detail.label,
-        timecode=detail.timecode,
-        notes=detail.notes,
-    )
-
-
-def _report_frame_details_for_frames(ctx: RunContext, *, frames: list[int]) -> list[FrameDetail]:
-    if ctx.selection_breakdown is None and ctx.selection_details_by_source_frame is None:
-        return []
-
-    frame_details: list[FrameDetail] = []
-    for aligned_frame in frames:
-        source_frame = _map_aligned_to_source_frame(
-            clip=ctx.reference,
-            aligned_frame=aligned_frame,
-        )
-        selection_detail = _selection_detail_for_frame(
-            source_frame,
-            ctx.selection_details_by_source_frame,
-        )
-        selection_label = (
-            selection_detail.label
-            if selection_detail is not None
-            else selection_label_for_frame(source_frame, ctx.selection_breakdown)
-        )
-        frame_details.append(
-            frame_detail_for_source_frame(
-                source_frame=source_frame,
-                selection_detail=_report_selection_detail(selection_detail),
-                selection_label=selection_label,
-            )
-        )
-    return frame_details
-
-
-def _selection_breakdown_with_source_offset(
-    breakdown: SelectionBreakdown,
-    *,
-    source_offset: int,
-) -> SelectionBreakdown:
-    return SelectionBreakdown(
-        user=[frame + source_offset for frame in breakdown.user],
-        quantile_dark=[frame + source_offset for frame in breakdown.quantile_dark],
-        quantile_bright=[frame + source_offset for frame in breakdown.quantile_bright],
-        motion=[frame + source_offset for frame in breakdown.motion],
-        random=[frame + source_offset for frame in breakdown.random],
-    )
-
-
-def _selection_details_with_source_offset(
-    details_by_frame: SelectionDetailsByFrame,
-    *,
-    source_offset: int,
-    fps: Fraction,
-) -> SelectionDetailsByFrame:
-    shifted_details: SelectionDetailsByFrame = {}
-    for frame_index, detail in details_by_frame.items():
-        source_frame = frame_index + source_offset
-        shifted_details[source_frame] = SelectionDetail(
-            frame_index=source_frame,
-            label=detail.label,
-            source=detail.source,
-            timecode=_selection_timecode_for_frame(source_frame, fps),
-            score=detail.score,
-            clip_role=detail.clip_role,
-            notes=detail.notes,
-        )
-    return shifted_details
-
-
-def _selection_timecode_for_frame(frame_index: int, fps: Fraction) -> str | None:
-    if fps <= 0:
-        return None
-    total_milliseconds = round((Fraction(frame_index, 1) * 1000) / fps)
-    total_seconds, milliseconds = divmod(total_milliseconds, 1000)
-    total_minutes, seconds = divmod(total_seconds, 60)
-    hours, minutes = divmod(total_minutes, 60)
-    return f"{hours:02}:{minutes:02}:{seconds:02}.{milliseconds:03}"
-
-
-def _to_overlay_selection_detail(detail: SelectionDetail) -> OverlaySelectionDetail:
-    from frame_compare.render.types import OverlaySelectionDetail
-
-    return OverlaySelectionDetail(
-        frame_index=detail.frame_index,
-        label=detail.label,
-        source=detail.source,
-        timecode=detail.timecode,
-        score=detail.score,
-        clip_role=detail.clip_role,
-        notes=detail.notes,
     )
 
 
@@ -826,7 +360,7 @@ def run_align_phase(ctx: RunContext, *, selected_frames: list[int]) -> AlignPhas
         for comp, comp_trim in zip(updated_comparisons, comp_trims, strict=True)
     ]
     selection_source_window = _selection_source_window_for_alignment(ctx)
-    selected_source_frames = _source_frames_for_reference_base_domain(
+    selected_source_frames = source_frames_for_reference_base_domain(
         reference=ctx.reference,
         selected_frames=selected_frames,
     )
@@ -836,9 +370,9 @@ def run_align_phase(ctx: RunContext, *, selected_frames: list[int]) -> AlignPhas
         reference=reference,
         comparisons=comparisons,
         selection_source_window=selection_source_window,
-        generated_requested_count=_generated_frame_count(ctx.config.analysis),
+        generated_requested_count=generated_frame_count(ctx.config.analysis),
         seed=ctx.config.analysis.random_seed,
-        allow_fallback=_generated_frame_count(ctx.config.analysis) > 0,
+        allow_fallback=generated_frame_count(ctx.config.analysis) > 0,
     )
     normalized_selected_frames = normalized_selection.selected_frames
     if normalized_selection.dropped_user_source_frames:
@@ -887,7 +421,7 @@ def run_align_phase(ctx: RunContext, *, selected_frames: list[int]) -> AlignPhas
                         frame_index=frame + reference.trim.trim_start_frames,
                         label="User",
                         source="frame_plan",
-                        timecode=_selection_timecode_for_frame(
+                        timecode=selection_timecode_for_frame(
                             frame + reference.trim.trim_start_frames,
                             reference.effective_fps,
                         ),
@@ -905,7 +439,7 @@ def run_align_phase(ctx: RunContext, *, selected_frames: list[int]) -> AlignPhas
             user=normalized_selection.user_source_frames,
             random=normalized_selection.generated_source_frames,
         )
-        selection_details_by_source_frame = _initial_selection_details_by_source_frame(
+        selection_details_by_source_frame = build_initial_selection_details_by_source_frame(
             user_source_frames=normalized_selection.user_source_frames,
             random_source_frames=normalized_selection.generated_source_frames,
             fps=reference.effective_fps,
@@ -918,12 +452,6 @@ def run_align_phase(ctx: RunContext, *, selected_frames: list[int]) -> AlignPhas
         selection_details_by_source_frame=selection_details_by_source_frame,
         warnings=warnings,
     )
-
-
-def _source_frames_for_reference_base_domain(
-    *, reference: ClipState, selected_frames: list[int]
-) -> list[int]:
-    return [reference.trim.trim_start_frames + frame for frame in selected_frames]
 
 
 def _selection_source_window_for_alignment(ctx: RunContext) -> tuple[int, int] | None:
@@ -970,13 +498,13 @@ def run_render_phase(
 
     overlay_mode = ctx.config.screenshots.overlay_mode
     reference_source_frames = [
-        _map_aligned_to_source_frame(clip=ctx.reference, aligned_frame=aligned_frame)
+        map_aligned_to_source_frame(clip=ctx.reference, aligned_frame=aligned_frame)
         for aligned_frame in frames
     ]
     selection_details: list[OverlaySelectionDetail | None] = [
-        _to_overlay_selection_detail(detail)
+        to_overlay_selection_detail(detail)
         if (
-            detail := _selection_detail_for_frame(
+            detail := selection_detail_for_frame(
                 source_frame,
                 ctx.selection_details_by_source_frame,
             )
@@ -995,7 +523,7 @@ def run_render_phase(
     batch_requests: list[ScreenshotBatchRequest] = []
     for clip in clips_state:
         source_frames = [
-            _map_aligned_to_source_frame(clip=clip, aligned_frame=aligned_frame)
+            map_aligned_to_source_frame(clip=clip, aligned_frame=aligned_frame)
             for aligned_frame in frames
         ]
         diagnostic_metadata = [
@@ -1057,203 +585,6 @@ def run_render_phase(
     )
 
 
-async def run_metadata_phase(
-    ctx: RunContext,
-    *,
-    client: httpx.AsyncClient | None,
-    metadata_prefetch: MetadataPrefetch,
-) -> MetadataPhaseOutput:
-    if metadata_prefetch.was_attempted:
-        return MetadataPhaseOutput(resolved_metadata=metadata_prefetch.metadata)
-
-    if client is None or not ctx.config.tmdb.enabled:
-        return MetadataPhaseOutput(resolved_metadata=None)
-    metadata = await resolve_run_metadata(
-        filenames=[ctx.reference.path.name],
-        config=ctx.config,
-        client=client,
-    )
-    return MetadataPhaseOutput(resolved_metadata=metadata)
-
-
-def record_dovi_not_implemented_warning(_ctx: RunContext) -> DoviPhaseOutput:
-    return DoviPhaseOutput(
-        warning="dovi: DOVI processing is not implemented yet; continuing without Dolby Vision extraction."
-    )
-
-
-async def run_publish_phase(
-    ctx: RunContext,
-    *,
-    client: httpx.AsyncClient | None,
-    metadata: TmdbMetadata | None,
-    render: RenderArtifacts | None = None,
-    selected_frames: list[int] | None = None,
-) -> PublishPhaseOutput:
-    if client is None:
-        return PublishPhaseOutput(slowpics_url=None)
-    if render is None:
-        raise SlowpicsError("No render artifacts available for slow.pics upload")
-    if selected_frames is None:
-        raise SlowpicsError("No selected frames available for slow.pics upload")
-
-    upload_plan = build_slowpics_upload_plan(
-        selected_frames=selected_frames,
-        clips=_slowpics_upload_clips(ctx),
-        screenshots_by_label=render.screenshots_by_label,
-    )
-    screenshot_dir = (
-        render.screenshot_dir
-        if render.screenshot_dir is not None
-        else ctx.workspace.screenshots_dir
-    )
-    result = await publish_to_slowpics(
-        screenshot_dir=screenshot_dir,
-        config=ctx.config.slowpics,
-        client=client,
-        metadata=metadata,
-        progress=ctx.reporter,
-        upload_plan=upload_plan,
-    )
-    post_upload_actions = await run_slowpics_post_upload_actions(
-        SlowpicsPostUploadRequest(
-            workspace=ctx.workspace,
-            config=ctx.config.slowpics,
-            slowpics_url=result.url,
-            metadata_title=metadata.title if metadata is not None else None,
-            upload_title=screenshot_dir.name,
-        )
-    )
-    return PublishPhaseOutput(
-        slowpics_url=result.url,
-        uploaded_file_paths=result.uploaded_file_paths,
-        post_upload_actions=post_upload_actions,
-    )
-
-
-def run_confirm_slowpics_upload_phase(
-    _ctx: RunContext,
-    *,
-    report_path: Path | None,
-    report_succeeded: bool,
-    confirm_slowpics_upload: SlowpicsUploadConfirmationFn | None,
-) -> ConfirmSlowpicsUploadPhaseOutput:
-    if not report_succeeded or report_path is None:
-        return ConfirmSlowpicsUploadPhaseOutput(
-            status="report_unavailable",
-            warnings=[REPORT_CONFIRMATION_UNAVAILABLE_WARNING],
-        )
-    if confirm_slowpics_upload is None:
-        validation_errors: list[dict[str, JSONValue]] = [
-            {
-                "type": "value_error",
-                "loc": ["slowpics", "confirm_upload_after_report"],
-                "msg": "Report-confirmed slow.pics upload requires a confirmation callback.",
-                "input": True,
-            }
-        ]
-        raise ConfigValidationError(
-            validation_errors,
-            message="Report-confirmed slow.pics upload requires a confirmation callback",
-            hint=(
-                "Provide RunDependencies.confirm_slowpics_upload, or disable "
-                "slowpics.confirm_upload_after_report."
-            ),
-        )
-
-    decision = confirm_slowpics_upload(SlowpicsUploadConfirmationRequest(report_path=report_path))
-    return ConfirmSlowpicsUploadPhaseOutput(status=decision)
-
-
-def _slowpics_upload_clips(ctx: RunContext) -> list[SlowpicsUploadClip]:
-    clips = [ctx.reference, *ctx.comparisons]
-    seen_labels: set[str] = set()
-    upload_clips: list[SlowpicsUploadClip] = []
-    for clip in clips:
-        if clip.label in seen_labels:
-            raise SlowpicsError(f"Duplicate clip label in slow.pics upload input: {clip.label!r}")
-        seen_labels.add(clip.label)
-        upload_clips.append(SlowpicsUploadClip(label=clip.label, image_name=clip.path.stem))
-    return upload_clips
-
-
-def run_report_phase(
-    ctx: RunContext,
-    *,
-    frames: list[int],
-    render: RenderArtifacts | None,
-    metadata: TmdbMetadata | None,
-    slowpics_url: str | None,
-) -> ReportPhaseOutput:
-    if render is None or not render.screenshots_by_label:
-        return ReportPhaseOutput(report_path=None, report_succeeded=True)
-
-    clips = [ctx.reference, *ctx.comparisons]
-    clip_info = [
-        clip_info_from_state(clip, render.screenshots_by_label[clip.label]) for clip in clips
-    ]
-    report_data = ReportData(
-        clips=clip_info,
-        frames=frames,
-        metadata=metadata,
-        slowpics_url=slowpics_url,
-        frame_details=_report_frame_details_for_frames(ctx, frames=frames),
-    )
-    report_path = generate_report(report_data, ctx.config.report)
-    return ReportPhaseOutput(report_path=report_path, report_succeeded=True)
-
-
-def run_post_report_cleanup_phase(
-    ctx: RunContext,
-    *,
-    uploaded_file_paths: tuple[Path, ...],
-    report_succeeded: bool,
-) -> PostReportCleanupPhaseOutput:
-    if not _should_delete_uploaded_files_after_report(
-        ctx,
-        uploaded_file_paths=uploaded_file_paths,
-        report_succeeded=report_succeeded,
-    ):
-        return PostReportCleanupPhaseOutput()
-
-    warnings: list[str] = []
-    deleted_count = 0
-    for path in uploaded_file_paths:
-        try:
-            path.unlink()
-            deleted_count += 1
-        except OSError as exc:
-            message = f"cleanup: failed to delete uploaded screenshot {path}: {exc}"
-            warnings.append(message)
-            log.warning(
-                "slowpics_uploaded_file_delete_failed",
-                path=str(path),
-                error=str(exc),
-            )
-
-    if deleted_count:
-        log.info("slowpics_uploaded_files_deleted", count=deleted_count)
-
-    return PostReportCleanupPhaseOutput(warnings=warnings)
-
-
-def _should_delete_uploaded_files_after_report(
-    ctx: RunContext,
-    *,
-    uploaded_file_paths: tuple[Path, ...],
-    report_succeeded: bool,
-) -> bool:
-    if not ctx.config.slowpics.delete_after_upload:
-        return False
-    if not uploaded_file_paths:
-        return False
-    if not ctx.config.report.enable:
-        return True
-    if not report_succeeded:
-        return False
-    return ctx.config.report.embed_images
-
-
 @dataclass(frozen=True)
 class _TrimmedOverlapSelection:
     selected_frames: list[int]
@@ -1298,7 +629,7 @@ def _reselect_frames_for_trimmed_overlap(
         comparisons=comparisons,
         selection_source_window=selection_source_window,
     )
-    target_count = _generated_frame_count(config)
+    target_count = generated_frame_count(config)
     if selectable_length <= 0 or target_count <= 0:
         return None
     if target_count > selectable_length:
@@ -1333,11 +664,11 @@ def _reselect_frames_for_trimmed_overlap(
             for frame in trimmed_selection.frames
             if frame not in local_user_frame_set
         ],
-        selection_breakdown=_selection_breakdown_with_source_offset(
+        selection_breakdown=selection_breakdown_with_source_offset(
             trimmed_selection.breakdown,
             source_offset=selectable_start,
         ),
-        selection_details_by_source_frame=_selection_details_with_source_offset(
+        selection_details_by_source_frame=selection_details_with_source_offset(
             dict(trimmed_selection.selection_details),
             source_offset=selectable_start,
             fps=trimmed_metrics.metadata.fps,
@@ -1538,17 +869,3 @@ def _selectable_aligned_source_window(
         overlap_start = max(overlap_start, selection_start)
         overlap_end = min(overlap_end, selection_end)
     return overlap_start, max(0, overlap_end - overlap_start)
-
-
-def _map_aligned_to_source_frame(*, clip: ClipState, aligned_frame: int) -> int:
-    source_frame = clip.trim.trim_start_frames + aligned_frame
-    trim_end = (
-        clip.trim.trim_end_frame_inclusive
-        if clip.trim.trim_end_frame_inclusive is not None
-        else clip.probe.num_frames - 1
-    )
-    if source_frame > trim_end:
-        raise AudioAlignmentError(
-            f"Aligned frame {aligned_frame} exceeds trimmed domain for {clip.path.name}."
-        )
-    return source_frame
