+++ src/frame_compare/orchestration/preparation.py
@@ -10,7 +10,10 @@ import structlog
 import frame_compare.analysis.cache_io as cache_io
 from frame_compare.analysis.errors import MetricsCalculationError
 from frame_compare.analysis.window import SelectionWindow
-from frame_compare.config.overrides import apply_cli_overrides
+from frame_compare.config.effective import (
+    build_preflight_input_dir_override,
+    resolve_effective_config,
+)
 from frame_compare.config.schema import ConfigSchema
 from frame_compare.config.schema_models import SourceOverrideConfig
 from frame_compare.orchestration.analysis_policy import (
@@ -23,7 +26,7 @@ from frame_compare.orchestration.context import (
     ClipState,
 )
 from frame_compare.orchestration.errors import MixedSourceFpsError
-from frame_compare.orchestration.phase_tasks import resolve_run_metadata
+from frame_compare.orchestration.phase_post_render import resolve_run_metadata
 from frame_compare.orchestration.preflight import discover_inputs, prepare_preflight
 from frame_compare.orchestration.probing.probe_cache import (
     compute_probe_cache_key,
@@ -57,13 +60,6 @@ from frame_compare.utils.cache_errors import CacheCorruptionError, CacheVersionM
 from frame_compare.utils.types import WorkspacePaths
 
 log = structlog.get_logger()
-
-
-def _build_preflight_overrides(request: RunRequest) -> dict[str, object] | None:
-    overrides: dict[str, object] = {}
-    if request.input_dir is not None:
-        overrides["paths"] = {"input_dir": str(request.input_dir)}
-    return overrides or None
 
 
 def _remove_cached_metrics(
@@ -305,17 +301,14 @@ async def execute_prep(
     preflight = prepare_preflight(
         root=request.root,
         config_path=request.config_path,
-        overrides=_build_preflight_overrides(request),
+        overrides=build_preflight_input_dir_override(request.input_dir),
     )
     preflight_end = deps.clock()
     preflight_duration = (preflight_end - preflight_start).total_seconds()
 
     load_sources_start = deps.clock()
     workspace = preflight.workspace
-    config = apply_cli_overrides(
-        preflight.config,
-        cli_args=request.cli_config_overrides(),
-    )
+    config = resolve_effective_config(preflight.config, request.cli_config_overrides())
     validate_skip_analysis_frame_selection_contract(
         skip_analysis=request.skip_analysis,
         config=config.analysis,
