+++ tests/orchestration/test_phase_tasks.py
@@ -21,7 +21,7 @@ from frame_compare.analysis.types import (
 )
 from frame_compare.analysis.window import SelectionWindow
 from frame_compare.config.errors import ConfigValidationError
-from frame_compare.orchestration import phase_tasks
+from frame_compare.orchestration import phase_post_render, phase_selection
 from frame_compare.orchestration.types import (
     RenderArtifacts,
     RunArtifacts,
@@ -82,12 +82,12 @@ def test_run_analyze_phase_records_cache_hit_and_selection_breakdown(
         calls["select"] = kwargs
         return selection
 
-    monkeypatch.setattr(phase_tasks.cache_io, "load_cached_metrics", _fake_load_cached_metrics)
-    monkeypatch.setattr(phase_tasks, "calculate_metrics", _fake_calculate_metrics)
-    monkeypatch.setattr(phase_tasks, "select_frames", _fake_select_frames)
+    monkeypatch.setattr(phase_selection.cache_io, "load_cached_metrics", _fake_load_cached_metrics)
+    monkeypatch.setattr(phase_selection, "calculate_metrics", _fake_calculate_metrics)
+    monkeypatch.setattr(phase_selection, "select_frames", _fake_select_frames)
     selected_frames: list[int] = []
 
-    output = phase_tasks.run_analyze_phase(
+    output = phase_selection.run_analyze_phase(
         ctx,
         input_videos=input_videos,
         workspace=ctx.workspace,
@@ -148,12 +148,12 @@ def test_run_analyze_phase_uses_prepared_analysis_selection_domain(
             breakdown=SelectionBreakdown(quantile_dark=[0]),
         )
 
-    monkeypatch.setattr(phase_tasks.cache_io, "compute_cache_key", _fake_compute_cache_key)
-    monkeypatch.setattr(phase_tasks.cache_io, "load_cached_metrics", _fake_load_cached_metrics)
-    monkeypatch.setattr(phase_tasks, "calculate_metrics", _fake_calculate_metrics)
-    monkeypatch.setattr(phase_tasks, "select_frames", _fake_select_frames)
+    monkeypatch.setattr(phase_selection.cache_io, "compute_cache_key", _fake_compute_cache_key)
+    monkeypatch.setattr(phase_selection.cache_io, "load_cached_metrics", _fake_load_cached_metrics)
+    monkeypatch.setattr(phase_selection, "calculate_metrics", _fake_calculate_metrics)
+    monkeypatch.setattr(phase_selection, "select_frames", _fake_select_frames)
 
-    phase_tasks.run_analyze_phase(
+    phase_selection.run_analyze_phase(
         ctx,
         input_videos=input_videos,
         workspace=ctx.workspace,
@@ -177,11 +177,11 @@ def test_run_analyze_phase_cache_only_missing_cache_does_not_recompute(
     def _fake_calculate_metrics(**_kwargs: object) -> FrameMetrics:
         raise AssertionError("cache-only analyze phase must not recompute metrics")
 
-    monkeypatch.setattr(phase_tasks.cache_io, "load_cached_metrics", _fake_load_cached_metrics)
-    monkeypatch.setattr(phase_tasks, "calculate_metrics", _fake_calculate_metrics)
+    monkeypatch.setattr(phase_selection.cache_io, "load_cached_metrics", _fake_load_cached_metrics)
+    monkeypatch.setattr(phase_selection, "calculate_metrics", _fake_calculate_metrics)
 
     with pytest.raises(MetricsCalculationError, match="Cached metrics missing"):
-        phase_tasks.run_analyze_phase(
+        phase_selection.run_analyze_phase(
             ctx,
             input_videos=input_videos,
             workspace=ctx.workspace,
@@ -236,10 +236,10 @@ def test_run_analyze_phase_selects_from_reference_base_trim_domain(
             },
         )
 
-    monkeypatch.setattr(phase_tasks.cache_io, "load_cached_metrics", _fake_load_cached_metrics)
-    monkeypatch.setattr(phase_tasks, "select_frames", _fake_select_frames)
+    monkeypatch.setattr(phase_selection.cache_io, "load_cached_metrics", _fake_load_cached_metrics)
+    monkeypatch.setattr(phase_selection, "select_frames", _fake_select_frames)
 
-    output = phase_tasks.run_analyze_phase(
+    output = phase_selection.run_analyze_phase(
         ctx,
         input_videos=input_videos,
         workspace=ctx.workspace,
@@ -292,10 +292,10 @@ def test_run_analyze_phase_selects_from_global_selection_window(
             },
         )
 
-    monkeypatch.setattr(phase_tasks.cache_io, "load_cached_metrics", _fake_load_cached_metrics)
-    monkeypatch.setattr(phase_tasks, "select_frames", _fake_select_frames)
+    monkeypatch.setattr(phase_selection.cache_io, "load_cached_metrics", _fake_load_cached_metrics)
+    monkeypatch.setattr(phase_selection, "select_frames", _fake_select_frames)
 
-    output = phase_tasks.run_analyze_phase(
+    output = phase_selection.run_analyze_phase(
         ctx,
         input_videos=input_videos,
         workspace=ctx.workspace,
@@ -316,7 +316,7 @@ def test_select_initial_frame_plan_uses_effective_selection_domain(tmp_path: Pat
     ctx.selection_window = SelectionWindow(start_frame=0, end_frame_exclusive=10)
     selected_frames: list[int] = []
 
-    output = phase_tasks.select_initial_frame_plan(ctx)
+    output = phase_selection.select_initial_frame_plan(ctx)
 
     assert selected_frames == []
     assert len(output.selected_frames) == 3
@@ -327,7 +327,7 @@ def test_select_initial_frame_plan_uses_global_selection_window(tmp_path: Path) 
     ctx = _context(tmp_path)
     ctx.selection_window = SelectionWindow(start_frame=24, end_frame_exclusive=48)
 
-    output = phase_tasks.select_initial_frame_plan(ctx)
+    output = phase_selection.select_initial_frame_plan(ctx)
 
     assert len(output.selected_frames) == 3
     assert all(24 <= frame < 48 for frame in output.selected_frames)
@@ -344,7 +344,7 @@ def test_select_initial_frame_plan_fails_when_user_random_candidates_are_empty(
     )
 
     with pytest.raises(SelectionError, match="no selectable user or random frames"):
-        phase_tasks.select_initial_frame_plan(ctx)
+        phase_selection.select_initial_frame_plan(ctx)
 
 
 def test_select_initial_frame_plan_warns_when_user_frames_are_dropped(
@@ -357,7 +357,7 @@ def test_select_initial_frame_plan_warns_when_user_frames_are_dropped(
         update={"user_frames": [0, 12], "random_frame_count": 1}
     )
 
-    output = phase_tasks.select_initial_frame_plan(ctx)
+    output = phase_selection.select_initial_frame_plan(ctx)
 
     assert 2 in output.selected_frames
     assert output.warnings == ["frame selection: dropped user frame(s) outside trims/windowing: 0"]
@@ -371,7 +371,7 @@ def test_select_initial_frame_plan_labels_user_and_random_frames(
         update={"user_frames": [0], "random_frame_count": 1}
     )
 
-    output = phase_tasks.select_initial_frame_plan(ctx)
+    output = phase_selection.select_initial_frame_plan(ctx)
 
     assert output.selection_breakdown.user == [0]
     assert len(output.selection_breakdown.random) == 1
@@ -387,7 +387,7 @@ def test_select_initial_frame_plan_refills_random_after_user_collision(tmp_path:
         update={"user_frames": [8], "random_frame_count": 10, "random_seed": 42}
     )
 
-    output = phase_tasks.select_initial_frame_plan(ctx)
+    output = phase_selection.select_initial_frame_plan(ctx)
 
     assert output.selection_breakdown.user == [8]
     assert len(output.selection_breakdown.random) == 10
@@ -405,7 +405,7 @@ def test_select_initial_frame_plan_fails_when_random_request_exceeds_remaining_d
     )
 
     with pytest.raises(SelectionError) as exc_info:
-        phase_tasks.select_initial_frame_plan(ctx)
+        phase_selection.select_initial_frame_plan(ctx)
 
     assert exc_info.value.context.details == {
         "reason": "insufficient random candidates after user frames",
@@ -440,7 +440,7 @@ def test_run_confirm_slowpics_upload_phase_marks_report_unavailable_without_prom
         callback_calls.append(request)
         return "confirmed"
 
-    output = phase_tasks.run_confirm_slowpics_upload_phase(
+    output = phase_post_render.run_confirm_slowpics_upload_phase(
         ctx,
         report_path=None,
         report_succeeded=False,
@@ -460,7 +460,7 @@ def test_run_confirm_slowpics_upload_phase_requires_callback_when_report_availab
     ctx = _context(tmp_path)
 
     with pytest.raises(ConfigValidationError, match="requires a confirmation callback"):
-        phase_tasks.run_confirm_slowpics_upload_phase(
+        phase_post_render.run_confirm_slowpics_upload_phase(
             ctx,
             report_path=tmp_path / "report.html",
             report_succeeded=True,
@@ -479,7 +479,7 @@ def test_run_confirm_slowpics_upload_phase_records_callback_decision(tmp_path: P
         callback_calls.append(request)
         return "declined"
 
-    output = phase_tasks.run_confirm_slowpics_upload_phase(
+    output = phase_post_render.run_confirm_slowpics_upload_phase(
         ctx,
         report_path=report_path,
         report_succeeded=True,
@@ -528,11 +528,11 @@ category_preference = "tv"
         captured["client"] = client
         return expected
 
-    monkeypatch.setattr(phase_tasks, "resolve_metadata", _fake_resolve_metadata)
+    monkeypatch.setattr(phase_post_render, "resolve_metadata", _fake_resolve_metadata)
 
     async def _run() -> TmdbMetadata | None:
         async with httpx.AsyncClient() as client:
-            result = await phase_tasks.resolve_run_metadata(
+            result = await phase_post_render.resolve_run_metadata(
                 filenames=["Heat.1995.mkv"],
                 config=config,
                 client=client,
