+++ tests/orchestration/test_phase_tasks_outputs.py
@@ -18,7 +18,7 @@ from frame_compare.analysis.types import (
     SelectionDetail,
 )
 from frame_compare.config.schema import OverlayMode
-from frame_compare.orchestration import phase_tasks
+from frame_compare.orchestration import phase_post_render, phase_tasks
 from frame_compare.orchestration.context import ClipActiveRect
 from frame_compare.orchestration.execution import build_phases_after_align
 from frame_compare.orchestration.phases import execute_phases
@@ -66,10 +66,10 @@ async def test_run_metadata_phase_resolves_when_enabled_and_client_present(
         captured.update(kwargs)
         return expected
 
-    monkeypatch.setattr(phase_tasks, "resolve_run_metadata", _fake_resolve_run_metadata)
+    monkeypatch.setattr(phase_post_render, "resolve_run_metadata", _fake_resolve_run_metadata)
 
     async with httpx.AsyncClient() as client:
-        output = await phase_tasks.run_metadata_phase(
+        output = await phase_post_render.run_metadata_phase(
             ctx,
             client=client,
             metadata_prefetch=MetadataPrefetch(None, False),
@@ -105,9 +105,9 @@ def test_run_report_phase_builds_report_data_and_records_path(
         captured["report_config"] = report_config
         return expected_path
 
-    monkeypatch.setattr(phase_tasks, "generate_report", _fake_generate_report)
+    monkeypatch.setattr(phase_post_render, "generate_report", _fake_generate_report)
 
-    output = phase_tasks.run_report_phase(
+    output = phase_post_render.run_report_phase(
         ctx,
         frames=[5],
         render=artifacts.render,
@@ -154,9 +154,9 @@ def test_run_report_phase_builds_four_clip_payload_inputs_in_clip_order(
         captured["report_config"] = report_config
         return tmp_path / "report.html"
 
-    monkeypatch.setattr(phase_tasks, "generate_report", _fake_generate_report)
+    monkeypatch.setattr(phase_post_render, "generate_report", _fake_generate_report)
 
-    phase_tasks.run_report_phase(
+    phase_post_render.run_report_phase(
         ctx,
         frames=[12],
         render=render,
@@ -219,9 +219,9 @@ def test_run_report_phase_passes_reference_source_frame_details(
         captured["report_config"] = report_config
         return expected_path
 
-    monkeypatch.setattr(phase_tasks, "generate_report", _fake_generate_report)
+    monkeypatch.setattr(phase_post_render, "generate_report", _fake_generate_report)
 
-    output = phase_tasks.run_report_phase(
+    output = phase_post_render.run_report_phase(
         ctx,
         frames=[1, 2],
         render=render,
@@ -715,7 +715,7 @@ def test_run_report_phase_labels_skipped_analysis_alignment_fallback_random_fram
         return expected_path
 
     monkeypatch.setattr(phase_tasks, "align_clips", _fake_align_clips)
-    monkeypatch.setattr(phase_tasks, "generate_report", _fake_generate_report)
+    monkeypatch.setattr(phase_post_render, "generate_report", _fake_generate_report)
 
     align_output = phase_tasks.run_align_phase(ctx, selected_frames=[0, 66])
     ctx.reference = align_output.reference
@@ -730,7 +730,7 @@ def test_run_report_phase_labels_skipped_analysis_alignment_fallback_random_fram
         screenshot_dir=tmp_path / "screenshots",
     )
 
-    output = phase_tasks.run_report_phase(
+    output = phase_post_render.run_report_phase(
         ctx,
         frames=align_output.selected_frames,
         render=render,
@@ -895,7 +895,7 @@ def test_output_phases_use_reselected_metric_metadata_after_real_initial_selecti
         "frame_compare.render.batch.orchestrator.render_screenshots_from_batch",
         _fake_render_screenshots_from_batch,
     )
-    monkeypatch.setattr(phase_tasks, "generate_report", _fake_generate_report)
+    monkeypatch.setattr(phase_post_render, "generate_report", _fake_generate_report)
 
     align_output = phase_tasks.run_align_phase(
         ctx,
@@ -924,7 +924,7 @@ def test_output_phases_use_reselected_metric_metadata_after_real_initial_selecti
         },
         screenshot_dir=tmp_path / "screenshots",
     )
-    report_output = phase_tasks.run_report_phase(
+    report_output = phase_post_render.run_report_phase(
         ctx,
         frames=align_output.selected_frames,
         render=render,
@@ -949,7 +949,7 @@ def test_run_report_phase_without_screenshots_clears_existing_report_path(tmp_pa
     ctx = _context(tmp_path)
     artifacts = RunArtifacts(report_path=tmp_path / "stale.html")
 
-    output = phase_tasks.run_report_phase(
+    output = phase_post_render.run_report_phase(
         ctx,
         frames=[1],
         render=artifacts.render,
@@ -992,7 +992,7 @@ async def test_run_publish_phase_sets_url_from_publish_result_and_delegates_post
     captured: dict[str, Any] = {}
     captured_clip_names: list[tuple[str, str]] = []
     captured_post_upload_request: SlowpicsPostUploadRequest | None = None
-    real_build_slowpics_upload_plan = phase_tasks.build_slowpics_upload_plan
+    real_build_slowpics_upload_plan = phase_post_render.build_slowpics_upload_plan
 
     def _capturing_build_slowpics_upload_plan(**kwargs: object) -> object:
         clips = cast(list[Any], kwargs["clips"])
@@ -1028,17 +1028,19 @@ async def test_run_publish_phase_sets_url_from_publish_result_and_delegates_post
         )
 
     monkeypatch.setattr(
-        phase_tasks, "build_slowpics_upload_plan", _capturing_build_slowpics_upload_plan
+        phase_post_render,
+        "build_slowpics_upload_plan",
+        _capturing_build_slowpics_upload_plan,
     )
-    monkeypatch.setattr(phase_tasks, "publish_to_slowpics", _fake_publish_to_slowpics)
+    monkeypatch.setattr(phase_post_render, "publish_to_slowpics", _fake_publish_to_slowpics)
     monkeypatch.setattr(
-        phase_tasks,
+        phase_post_render,
         "run_slowpics_post_upload_actions",
         _fake_run_slowpics_post_upload_actions,
     )
 
     async with httpx.AsyncClient() as client:
-        output = await phase_tasks.run_publish_phase(
+        output = await phase_post_render.run_publish_phase(
             ctx,
             client=client,
             metadata=metadata,
@@ -1095,7 +1097,7 @@ async def test_run_publish_phase_rejects_duplicate_clip_labels_at_translation_se
 
     async with httpx.AsyncClient() as client:
         with pytest.raises(SlowpicsError, match="Duplicate clip label in slow.pics upload input"):
-            await phase_tasks.run_publish_phase(
+            await phase_post_render.run_publish_phase(
                 ctx,
                 client=client,
                 metadata=None,
@@ -1132,15 +1134,15 @@ async def test_run_publish_phase_skips_shortcut_when_config_disabled(
     ) -> tuple[PostUploadActionResult, ...]:
         return ()
 
-    monkeypatch.setattr(phase_tasks, "publish_to_slowpics", _fake_publish_to_slowpics)
+    monkeypatch.setattr(phase_post_render, "publish_to_slowpics", _fake_publish_to_slowpics)
     monkeypatch.setattr(
-        phase_tasks,
+        phase_post_render,
         "run_slowpics_post_upload_actions",
         _no_post_upload_actions,
     )
 
     async with httpx.AsyncClient() as client:
-        output = await phase_tasks.run_publish_phase(
+        output = await phase_post_render.run_publish_phase(
             ctx,
             client=client,
             metadata=None,
@@ -1203,10 +1205,10 @@ async def test_unresolved_comparison_remains_in_render_report_and_slowpics_membe
         "frame_compare.render.batch.orchestrator.render_screenshots_from_batch",
         _fake_render_screenshots_from_batch,
     )
-    monkeypatch.setattr(phase_tasks, "generate_report", _fake_generate_report)
-    monkeypatch.setattr(phase_tasks, "publish_to_slowpics", _fake_publish_to_slowpics)
+    monkeypatch.setattr(phase_post_render, "generate_report", _fake_generate_report)
+    monkeypatch.setattr(phase_post_render, "publish_to_slowpics", _fake_publish_to_slowpics)
     monkeypatch.setattr(
-        phase_tasks,
+        phase_post_render,
         "run_slowpics_post_upload_actions",
         _no_post_upload_actions,
     )
@@ -1216,7 +1218,7 @@ async def test_unresolved_comparison_remains_in_render_report_and_slowpics_membe
         frames=frames,
         runner=cast(Any, _RenderRunner()),
     )
-    phase_tasks.run_report_phase(
+    phase_post_render.run_report_phase(
         ctx,
         frames=frames,
         render=render_output.render,
@@ -1224,7 +1226,7 @@ async def test_unresolved_comparison_remains_in_render_report_and_slowpics_membe
         slowpics_url=None,
     )
     async with httpx.AsyncClient() as client:
-        await phase_tasks.run_publish_phase(
+        await phase_post_render.run_publish_phase(
             ctx,
             client=client,
             metadata=None,
@@ -1448,7 +1450,7 @@ def test_post_report_cleanup_skips_non_embedded_reports(tmp_path: Path) -> None:
         path.parent.mkdir(parents=True, exist_ok=True)
         path.write_bytes(b"\x89PNG\r\n\x1a\n")
 
-    output = phase_tasks.run_post_report_cleanup_phase(
+    output = phase_post_render.run_post_report_cleanup_phase(
         ctx,
         uploaded_file_paths=uploaded,
         report_succeeded=True,
@@ -1477,7 +1479,7 @@ def test_post_report_cleanup_deletes_planned_files_after_embedded_report_success
         path.write_bytes(b"\x89PNG\r\n\x1a\n")
     shortcut.write_text("[InternetShortcut]\nURL=https://slow.pics/c/example\n", encoding="utf-8")
 
-    output = phase_tasks.run_post_report_cleanup_phase(
+    output = phase_post_render.run_post_report_cleanup_phase(
         ctx,
         uploaded_file_paths=uploaded,
         report_succeeded=True,
@@ -1502,7 +1504,7 @@ def test_post_report_cleanup_deletes_planned_files_when_reports_disabled(
         path.parent.mkdir(parents=True, exist_ok=True)
         path.write_bytes(b"\x89PNG\r\n\x1a\n")
 
-    output = phase_tasks.run_post_report_cleanup_phase(
+    output = phase_post_render.run_post_report_cleanup_phase(
         ctx,
         uploaded_file_paths=uploaded,
         report_succeeded=False,
@@ -1521,7 +1523,7 @@ def test_post_report_cleanup_skips_without_upload_handoff(tmp_path: Path) -> Non
     stale.parent.mkdir(parents=True, exist_ok=True)
     stale.write_bytes(b"\x89PNG\r\n\x1a\n")
 
-    output = phase_tasks.run_post_report_cleanup_phase(
+    output = phase_post_render.run_post_report_cleanup_phase(
         ctx,
         uploaded_file_paths=(),
         report_succeeded=False,
@@ -1542,7 +1544,7 @@ def test_post_report_cleanup_requires_report_success_when_report_enabled(
     uploaded[0].parent.mkdir(parents=True, exist_ok=True)
     uploaded[0].write_bytes(b"\x89PNG\r\n\x1a\n")
 
-    output = phase_tasks.run_post_report_cleanup_phase(
+    output = phase_post_render.run_post_report_cleanup_phase(
         ctx,
         uploaded_file_paths=uploaded,
         report_succeeded=False,
@@ -1573,9 +1575,9 @@ def test_post_report_cleanup_returns_warning_and_logs_for_delete_error(
         warning_calls.append((event, kwargs))
 
     monkeypatch.setattr(Path, "unlink", _raise_permission_error)
-    monkeypatch.setattr(phase_tasks.log, "warning", _capture_warning)
+    monkeypatch.setattr(phase_post_render.log, "warning", _capture_warning)
 
-    output = phase_tasks.run_post_report_cleanup_phase(
+    output = phase_post_render.run_post_report_cleanup_phase(
         ctx,
         uploaded_file_paths=uploaded,
         report_succeeded=False,
